internalize account checker and prio fee getter
