+++ examples/oft-solana/tasks/solana/index.ts
@@ -29,13 +29,12 @@ import { getSimulationComputeUnits } from '@solana-developers/helpers'
 import { backOff } from 'exponential-backoff'
 
 import { formatEid } from '@layerzerolabs/devtools'
-import { getSolanaKeypair } from '@layerzerolabs/devtools-solana'
+import { getPrioritizationFees, getSolanaKeypair } from '@layerzerolabs/devtools-solana'
 import { promptToContinue } from '@layerzerolabs/io-devtools'
 import { EndpointId, endpointIdToNetwork } from '@layerzerolabs/lz-definitions'
 import { OftPDA } from '@layerzerolabs/oft-v2-solana-sdk'
 
 import { DebugLogger, KnownWarnings, createSolanaConnectionFactory } from '../common/utils'
-import getFee from '../utils/getFee'
 
 const LOOKUP_TABLE_ADDRESS: Partial<Record<EndpointId, PublicKey>> = {
     [EndpointId.SOLANA_V2_MAINNET]: publicKey('AokBxha6VMLLgf97B5VYHEtqztamWmYERBmmFvjuTzJB'),
@@ -234,7 +233,7 @@ export const getComputeUnitPriceAndLimit = async (
     lookupTableAccount: AddressLookupTableAccount,
     transactionType: TransactionType
 ) => {
-    const { averageFeeExcludingZeros } = await getFee(connection)
+    const { averageFeeExcludingZeros } = await getPrioritizationFees(connection)
     const priorityFee = Math.round(averageFeeExcludingZeros)
     const computeUnitPrice = BigInt(priorityFee)
