+++ examples/oft-solana/tasks/solana/utils.ts
@@ -1,5 +1,4 @@
-import { Connection, PublicKey } from '@solana/web3.js'
-import { backOff } from 'exponential-backoff'
+import { Connection } from '@solana/web3.js'
 import { HardhatRuntimeEnvironment } from 'hardhat/types'
 
 import { ChainType, EndpointId, endpointIdToChainType } from '@layerzerolabs/lz-definitions'
@@ -10,30 +9,6 @@ import {
     SubtaskLoadConfigTaskArgs,
     TASK_LZ_OAPP_CONFIG_GET,
 } from '@layerzerolabs/ua-devtools-evm-hardhat'
-
-/**
- * Assert that the account is initialized on the Solana blockchain.  Due to eventual consistency, there is a race
- * between account creation and initialization.  This function will retry 10 times with backoff to ensure the account is
- * initialized.
- * @param connection {Connection}
- * @param publicKey {PublicKey}
- */
-export const assertAccountInitialized = async (connection: Connection, publicKey: PublicKey) => {
-    return backOff(
-        async () => {
-            const accountInfo = await connection.getAccountInfo(publicKey)
-            if (!accountInfo) {
-                throw new Error('Multisig account not found')
-            }
-            return accountInfo
-        },
-        {
-            maxDelay: 30000,
-            numOfAttempts: 10,
-            startingDelay: 5000,
-        }
-    )
-}
 
 export const findSolanaEndpointIdInGraph = async (
     hre: HardhatRuntimeEnvironment,
