+++ packages/devtools-solana/src/common/accounts.ts
@@ -1,6 +1,7 @@
 import { Factory, mapError, OmniAddress } from '@layerzerolabs/devtools'
 import { createModuleLogger, Logger, printJson } from '@layerzerolabs/io-devtools'
 import { AccountInfo, Connection, PublicKey } from '@solana/web3.js'
+import { backOff } from 'exponential-backoff'
 
 export type GetAccountInfo = Factory<[OmniAddress], AccountInfo<Buffer> | undefined>
 
@@ -16,3 +17,20 @@ export const createGetAccountInfo =
 
         return logger.debug(`Got account info for ${account}:\n\n${printJson(accountInfo)}`), accountInfo ?? undefined
     }
+
+/**
+ * Ensures a Solana account exists by retrying the lookup a few times.
+ *
+ * This is useful right after creating an account where propagation may be slow.
+ */
+export const assertAccountInitialized = async (connection: Connection, publicKey: PublicKey) =>
+    backOff(
+        async () => {
+            const accountInfo = await connection.getAccountInfo(publicKey)
+            if (!accountInfo) {
+                throw new Error('Account not found')
+            }
+            return accountInfo
+        },
+        { maxDelay: 30000, numOfAttempts: 10, startingDelay: 5000 }
+    )
