+++ packages/devtools-solana/src/transactions/index.ts
@@ -1,2 +1,3 @@
 export * from './serde'
 export * from './signer'
+export * from './fees'
